requirements.txt
astroid==2.2.5
Babel==2.6.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django>2.2.1, <3.0
django-cors-headers==3.0.1
django-colorfield==0.3.2
django-debug-toolbar==1.11
django-extensions==2.1.7
django-phonenumber-field==2.4.0
django-webpack-loader==0.6.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-serializer-extensions==1.0.0
easy-thumbnails==2.6
gunicorn==19.9.0
hashids==1.2.0
idna==2.8
isort==4.3.20
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
phonenumbers==8.10.12
Pillow==7.2.0
PyJWT==1.7.1
pylint==2.3.1
python-decouple==3.1
python-stdnum==1.11
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.2
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.25.3
whitenoise==4.1.2
wrapt==1.11.2
django-reset-migrations==0.3.1
django-notifications-hq==1.5.0
django-auditlog==0.4.5
suds-py3==1.3.3.0
common-tags==1.0.3
xhtml2pdf==0.2.3
xmltodict==0.12.0
django-test-pep8==0.1
flake8
factory-boy==2.12.0
ipython
channels==2.2.0
channels-redis==2.4.0
xlwt==1.3.0
psycopg2-binary
dj_database_url==0.5.0
anyjson==0.3.3
